requirements.txt
Flask==2.2.2
gunicorn
Werkzeug==2.2.2
torch==2.2.1
torchvision==0.17.1
pillow==10.2.0